inheritanse this way Blueprint --> DisconnectedImplementation --> ConnecctedImplementation
